feat: add constant and network_volume
