+++ src/tetra_rp/client.py
@@ -53,7 +53,7 @@ def remote(
             if mount_volume:
                 try:
                     network_volume = await mount_volume.deploy()
-                    resource_config.networkVolumeId = network_volume
+                    resource_config.networkVolumeId = network_volume.id
                     log.info(
                         f"Updated resource config with network volume: {network_volume}"
                     )
