+++ src/tetra_rp/core/resources/constants.py
@@ -1,0 +1,5 @@
+
+import os
+
+CONSOLE_BASE_URL = os.environ.get("CONSOLE_BASE_URL", "https://console.runpod.io")
+CONSOLE_URL = f"{CONSOLE_BASE_URL}/serverless/user/endpoint/%s"
